requirements.txt
Flask
flask-cors
python-dotenv
cryptography
requests
requests_ntlm
beautifulsoup4
selectolax
lxml
redis
pycurl
httpx
httpx-ntlm
PyPDF2
pdfminer.six
python-docx
python-pptx
